requirements.txt
accelerate==1.12.0
addict==2.4.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
blosc2==3.12.2
brotli==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.3
cycler==0.12.1
Cython==3.2.4
dash==3.3.0
decorator==5.2.1
diff_gaussian_rasterization @ file:///mnt/e/3D/experiments/gaussian-splatting/submodules/diff-gaussian-rasterization
diffusers==0.36.0
docstring_parser==0.17.0
einops==0.8.1
embreex==2.17.7.post7
executing==2.2.1
fastapi==0.128.0
fastjsonschema==2.21.2
ffmpy==1.0.0
filelock==3.20.3
Flask==3.1.2
fonttools==4.61.1
fsspec==2026.1.0
fused_ssim @ file:///mnt/e/3D/experiments/gaussian-splatting/submodules/fused-ssim
gitdb==4.0.12
GitPython==3.1.46
gradio==6.3.0
gradio_client==2.0.3
groovy==0.1.2
gsplat==1.5.3
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
iopath==0.1.10
ipython==9.9.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
itsdangerous==2.2.0
jaxtyping==0.3.5
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lazy_loader==0.4
lxml==6.0.2
manifold3d==3.3.2
mapbox_earcut==2.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
narwhals==2.15.0
nbformat==5.10.4
ndindex==1.10.1
nest-asyncio==1.6.0
networkx==3.6.1
ninja==1.13.0
nodeenv==1.10.0
numexpr==2.14.1
numpy==2.2.6
# nunchaku @ https://github.com/mit-han-lab/nunchaku/releases/download/v0.2.0/nunchaku-0.2.0+torch2.7-cp311-cp311-linux_x86_64.whl#sha256=07f180d2cc215c947f13ca484d388bef050479a11e717345c98b7d43a31dfeee
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
open3d==0.19.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.18.1
pexpect==4.9.0
pillow==12.1.0
pillow_heif==1.1.1
platformdirs==4.5.1
plotly==6.5.2
plyfile==1.1.3
portalocker==3.2.0
prompt_toolkit==3.0.52
protobuf==6.33.4
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py360convert==1.0.4
pycollada==0.9.2
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.1
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-multipart==0.0.21
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@cbcae096a0b9b04f7c515d11bb4285a82e96b8d7
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
retrying==1.4.2
rich==14.2.0
rpds-py==0.30.0
rtree==1.4.1
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.49.0
shapely==2.1.2
sharp @ file:///mnt/e/3D/experiments/WorldGen/submodules/ml-sharp
shellingham==1.5.4
simple_knn @ file:///mnt/e/3D/experiments/gaussian-splatting/submodules/simple-knn
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
starlette==0.50.0
svg.path==7.0
sympy==1.14.0
tables==3.10.2
tabulate==0.9.0
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2026.1.14
timm==1.0.24
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.6
trimesh==4.11.0
triton==3.5.1
typeguard==4.4.4
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==1.0.5
tzdata==2025.3
# unik3d @ git+https://github.com/lpiccinelli-eth/UniK3D.git@29f7862f3c3ee79c53ef0153dfc901ce511df575
urllib3==2.6.3
uvicorn==0.40.0
vhacdx==0.0.10
# viser @ git+https://github.com/ZiYang-xie/viser.git@210f1e3c73de1fc89fc6f61ed4152a315bdf991c
wadler_lindig==0.1.7
wandb==0.24.0
wcwidth==0.2.14
websockets==16.0
Werkzeug==3.1.5
widgetsnbextension==4.0.15
worldgen @ file:///mnt/e/3D/experiments/WorldGen
xformers==0.0.33.post2
xxhash==3.6.0
yourdfpy==0.0.59
zipp==3.23.0
